feat : 투표 처리, 투표 취소, 만료된 게시판 status 변경

+++ src/main/java/caugarde/vote/repository/v2/interfaces/jpa/StudentJpaRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface StudentJpaRepository extends JpaRepository<Student, Integer> {
+public interface StudentJpaRepository extends JpaRepository<Student, Long> {
 
     Optional<Student> findByEmail(String email);
